+++ experimental/packages/opentelemetry-configuration/test/ConfigProvider.test.ts
@@ -809,6 +809,18 @@ describe('ConfigProvider', function () {
         resource: {
           attributes_list:
             'service.namespace=my-namespace,service.version=1.0.0',
+          attributes: [
+            {
+              name: 'service.namespace',
+              value: 'my-namespace',
+              type: 'string',
+            },
+            {
+              name: 'service.version',
+              value: '1.0.0',
+              type: 'string',
+            },
+          ],
         },
       };
       const configProvider = createConfigProvider();